+++ c2/migrations/20251026144632_add_agent_id_to_ct.sql
@@ -1,0 +1,6 @@
+-- Add migration script here
+ALTER TABLE completed_tasks
+    ADD COLUMN agent_id TEXT;
+
+ALTER TABLE completed_tasks
+    ADD COLUMN command_id INT;