+++ c2/src/db.rs
@@ -135,7 +135,7 @@ impl Db {
                 self.mark_task_completed(&task)
                     .await
                     .expect("Could not complete task");
-                self.add_completed_task(&task)
+                self.add_completed_task(&task, uid)
                     .await
                     .expect("Could not add task to completed");
             }
@@ -229,16 +229,20 @@ impl Db {
 
     /// Adds a completed task into the `completed_tasks` table which stores the results
     /// and metadata associated with completed task results, to be used by the client.
-    pub async fn add_completed_task(&self, task: &Task) -> Result<(), sqlx::Error> {
+    pub async fn add_completed_task(&self, task: &Task, agent_id: &str) -> Result<(), sqlx::Error> {
+        let cmd_id: u32 = task.command.into();
+
         let _ = sqlx::query(
             r#"
-            INSERT INTO completed_tasks (task_id, result, time_completed_ms)
-            VALUES ($1, $2, $3)
+            INSERT INTO completed_tasks (task_id, result, time_completed_ms, agent_id, command_id)
+            VALUES ($1, $2, $3, $4, $5)
         "#,
         )
         .bind(task.id)
         .bind(task.metadata.as_deref())
         .bind(task.completed_time)
+        .bind(agent_id)
+        .bind(cmd_id as i32)
         .execute(&self.pool)
         .await?;
 
@@ -546,5 +550,41 @@ impl Db {
         .await?;
 
         Ok(rows)
+    }
+
+    pub async fn get_agent_export_data(&self, uid: &str) -> Result<Option<Vec<Task>>, sqlx::Error> {
+        let rows = sqlx::query(
+            r#"
+            SELECT task_id, result, time_completed_ms, command_id
+            FROM completed_tasks
+            WHERE agent_id = $1"#,
+        )
+        .bind(uid)
+        .fetch_all(&self.pool)
+        .await?;
+
+        if rows.is_empty() {
+            return Ok(None);
+        }
+
+        let mut results = vec![];
+
+        for row in rows {
+            let task_id: i32 = row.try_get("task_id")?;
+            let metadata: Option<String> = row.try_get("result")?;
+            let completed_time: i64 = row.try_get("time_completed_ms")?;
+            let command_id: i32 = row.try_get("command_id")?;
+
+            let command = Command::from_u32(command_id as _);
+
+            results.push(Task {
+                id: task_id,
+                command,
+                completed_time,
+                metadata,
+            });
+        }
+
+        Ok(Some(results))
     }
 }